Continued work on unit tests for the Hamiltonians.
To prevent repetition in the unit tests, my new approach is going to be to test the exact numerical form of the general Hamiltonian against formula, then compare each approximated Hamiltonian to the general in the limit of their assumptions. This should mean that the general Hamiltonian is very thoroughly tested, and everything else takes advantage of this.