Change of coordinate system
I've changed everything to use a different coordinate system.
The new coordinate system is the one used in Gildemeister's thesis, and I think it has a much simpler interpretation than the one we used for the MRF paper.
As a result, I've moved everything over to it.

Theta corresponds to the vertical angle - it is pi at the bottom of the shell, NOT ZERO.
Gamma corresponds to z-axis rotation angle - x axis has gamma=0 or pi.
(it's similar to standard spherical polar coordinates)

The change has allowed me to very easily get the QuadrupoleSurfaceSampler up and running. To test everything is correct, I've looked at the calculated coupling strengths and compared them to what I expected for different polarisations.
So far it all checks out!

The derivation is attached as a mathematica notebook.